// src/common/enums/role.enum.ts
export enum Role {
  ADMIN = 'ADMIN',
  USER = 'USER',
  GERENTE_PLANEACION = 'GERENTE_PLANEACION',
  SAR = 'SAR',
  SAN = 'SAN',
  SUBGERENTE_PLANEACION = 'SUBGERENTE_PLANEACION',
  SAN_IMPORTADOS = 'SAN_IMPORTADOS',
  GERENTE_ABASTECIMIENTO_REGIONAL = 'GERENTE_ABASTECIMIENTO_REGIONAL',
  GERENTE_ABASTECIMIENTO_NACIONAL = 'GERENTE_ABASTECIMIENTO_NACIONAL',
}